refactor: migrate Slinkity package.json
