--- .npmignore
@@ -1,5 +1,0 @@
-example/
-src/
-docs/
-.vscode/
-*.tgz
