--- client.js
@@ -1,4 +1,0 @@
-export { default as eagerLoader } from './packages/slinkity/client/eagerLoader'
-export { default as lazyLoader } from './packages/slinkity/client/lazyLoader'
-export { default as renderer } from './packages/slinkity/client/renderer'
-export { toMountPointById } from './packages/slinkity/client/toMountPointById'
