--- index.js
@@ -1,1 +1,0 @@
-export { defineConfig } from './packages/slinkity/main/defineConfig'
