+++ packages/slinkity/client.js
@@ -1,0 +1,4 @@
+export { default as eagerLoader } from './client/eagerLoader'
+export { default as lazyLoader } from './client/lazyLoader'
+export { default as renderer } from './client/renderer'
+export { toMountPointById } from './client/toMountPointById'
