+++ packages/slinkity/defineConfig.js
