+++ packages/slinkity/index.d.ts
@@ -1,0 +1,2 @@
+export type { UserSlinkityConfig } from './lib/config/defineConfig'
+export { defineConfig } from './lib/config/defineConfig'
