+++ packages/slinkity/index.js
@@ -1,0 +1,5 @@
+const { defineConfig } = require('./defineConfig')
+
+module.exports = {
+  defineConfig,
+}
