+++ packages/slinkity/plugin/index.js
@@ -7,7 +7,7 @@
  *  layouts: string;
  * }} dir - paths to all significant directories, as specified in 11ty's "dir" documentation
  * @property {import('../cli/toViteSSR').ViteSSR | null} viteSSR - utility to import components as Node-friendly modules
- * @property {import('../main/defineConfig').UserSlinkityConfig} userSlinkityConfig - Slinkity config options (either from user config or defaults)
+ * @property {import('../defineConfig').UserSlinkityConfig} userSlinkityConfig - Slinkity config options (either from user config or defaults)
  * @property {import('browser-sync').Options} browserSyncOptions - Slinkity's own browser sync server for dev environments
  * @property {'dev' | 'prod'} environment - whether we want a dev server or a production build
  */
